requirements.txt
SQLAlchemy==1.4.45
alembic==1.8.1
pytest==6.2.5
flake8==3.9.2
pyflakes==2.3.0
pytest-flake8==1.0.7
pytest-flakes==4.0.5
venusian==3.0.0
psycopg2-binary==2.9.5
redis==4.4.0
filedepot==0.9.0
bcrypt==4.0.1
falcon==3.1.1
PyJWT==2.6.0
pytz==2022.7
marshmallow==3.19.0
gunicorn==20.1.0
pillow~=9.0.0